+++ src/Nova3diLab/DF2/LOD/SubObject.cs
@@ -10,18 +10,18 @@ namespace Nova3diLab.Model.Lod
         public int FacesCount { get; set; }
         public int NormalsCount { get; set; }
         public int CollisionVolumesCount { get; set; }
-        public decimal XMinimum { get; set; }
-        public decimal XMaximum { get; set; }
-        public decimal YMinimum { get; set; }
-        public decimal YMaximum { get; set; }
-        public decimal ZMinimum { get; set; }
-        public decimal ZMaximum { get; set; }
-        public decimal XMedian { get; set; }
-        public decimal YMedian { get; set; }
-        public decimal ZMedian { get; set; }
-        public decimal SphereRadius { get; set; }
-        public decimal CircleRadius { get; set; }
-        public decimal ZTotal => ZMaximum - ZMinimum;
+        public double XMinimum { get; set; }
+        public double XMaximum { get; set; }
+        public double YMinimum { get; set; }
+        public double YMaximum { get; set; }
+        public double ZMinimum { get; set; }
+        public double ZMaximum { get; set; }
+        public double XMedian { get; set; }
+        public double YMedian { get; set; }
+        public double ZMedian { get; set; }
+        public double SphereRadius { get; set; }
+        public double CircleRadius { get; set; }
+        public double ZTotal => ZMaximum - ZMinimum;
 
         public byte[] Serialize()
         {
